EnterpriseWIFI_SD_LP5012_PM_8ADCs
ESP32 connects to an Enterprise WIFI network, samples a Particle sensor, along with 8ADCs, storing data to an SD card and sending over WIFI to a Server. The ESP32 also shows a reading on an array of LEDs, green for when all good, going red if readings go too high.